feat : detected data processing service logic #20
* 개발에 반영된 의도는 다음과 같습니다.

* 모델이 보내는 모든 데이터를 보관 한다.
* 받은 데이터 중, 유효성을 판단하고, 유효한 데이터만 광고와 관련된 디비에 반영해준다.

* 테스트 코드 곧 추가하겠습니다.

+++ src/server/src/main/java/com/drm/server/service/MediaService.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Service;
 @Service
 @RequiredArgsConstructor
 public class MediaService {
-    public void updateMediaData(int mediaId, boolean interest){
+    public void updateMediaData(Long mediaId, boolean interest){
 
     }
 }
